refactor(headless): split up analytics-utils
https://coveord.atlassian.net/browse/KIT-2468
