+++ packages/headless/src/features/facets/category-facet-set/category-facet-set-analytics-actions.ts
@@ -5,11 +5,8 @@ import {
   requiredNonEmptyString,
   validatePayload,
 } from '../../../utils/validate-payload';
-import {
-  AnalyticsType,
-  makeAnalyticsAction,
-  SearchAction,
-} from '../../analytics/analytics-utils';
+import {AnalyticsType, SearchAction} from '../../analytics/analytics-utils';
+import {makeAnalyticsAction} from '../../analytics/search-analytics-utils';
 import {facetIdDefinition} from '../generic/facet-actions-validation';
 
 export interface LogCategoryFacetBreadcrumbActionCreatorPayload {
