+++ packages/headless/src/features/facets/category-facet-set/category-facet-set-insight-analytics-actions.ts
@@ -5,11 +5,11 @@ import {
   requiredNonEmptyString,
   validatePayload,
 } from '../../../utils/validate-payload';
+import {AnalyticsType} from '../../analytics/analytics-utils';
 import {
-  AnalyticsType,
   InsightAction,
   makeInsightAnalyticsAction,
-} from '../../analytics/analytics-utils';
+} from '../../analytics/insight-analytics-utils';
 import {getCaseContextAnalyticsMetadata} from '../../case-context/case-context-state';
 import {facetIdDefinition} from '../generic/facet-actions-validation';
 import {LogCategoryFacetBreadcrumbActionCreatorPayload} from './category-facet-set-analytics-actions';
