+++ packages/headless/src/features/facets/generic/facet-generic-analytics-actions.ts
@@ -1,8 +1,5 @@
-import {
-  AnalyticsType,
-  makeAnalyticsAction,
-  SearchAction,
-} from '../../analytics/analytics-utils';
+import {AnalyticsType, SearchAction} from '../../analytics/analytics-utils';
+import {makeAnalyticsAction} from '../../analytics/search-analytics-utils';
 
 export const logClearBreadcrumbs = (): SearchAction =>
   makeAnalyticsAction(
