+++ packages/headless/src/features/insight-search/insight-search-actions-loader.ts
@@ -2,7 +2,7 @@ import {AsyncThunkAction} from '@reduxjs/toolkit';
 import {AsyncThunkInsightOptions} from '../../api/service/insight/insight-api-client';
 import {searchReducer as search} from '../../features/search/search-slice';
 import {InsightEngine} from '../../insight.index';
-import {InsightAction} from '../analytics/analytics-utils';
+import {InsightAction} from '../analytics/insight-analytics-utils';
 import {
   FetchQuerySuggestionsActionCreatorPayload,
   FetchQuerySuggestionsThunkReturn,
