+++ packages/headless/src/features/product-listing/product-listing-analytics.ts
@@ -4,11 +4,11 @@ import {
   AnalyticsType,
   ClickAction,
   documentIdentifier,
-  makeAnalyticsAction,
   partialDocumentInformation,
   ProductListingAction,
   validateResultPayload,
 } from '../analytics/analytics-utils';
+import {makeAnalyticsAction} from '../analytics/search-analytics-utils';
 
 export const logProductListing = (): ProductListingAction =>
   makeAnalyticsAction(
