+++ packages/headless/src/features/query-suggest/query-suggest-analytics-actions.ts
@@ -1,10 +1,7 @@
 import type {OmniboxSuggestionsMetadata} from 'coveo.analytics/dist/definitions/searchPage/searchPageEvents';
 import {SearchAppState} from '../../state/search-app-state';
-import {
-  AnalyticsType,
-  makeAnalyticsAction,
-  SearchAction,
-} from '../analytics/analytics-utils';
+import {AnalyticsType, SearchAction} from '../analytics/analytics-utils';
+import {makeAnalyticsAction} from '../analytics/search-analytics-utils';
 
 export interface LogQuerySuggestionClickActionCreatorPayload {
   /**
