+++ packages/headless/src/features/result-actions/result-actions-insight-analytics-actions.ts
@@ -2,11 +2,13 @@ import {Result} from '../../insight.index';
 import {
   AnalyticsType,
   documentIdentifier,
-  InsightAction,
-  makeInsightAnalyticsAction,
   partialDocumentInformation,
   validateResultPayload,
 } from '../analytics/analytics-utils';
+import {
+  InsightAction,
+  makeInsightAnalyticsAction,
+} from '../analytics/insight-analytics-utils';
 import {getCaseContextAnalyticsMetadata} from '../case-context/case-context-state';
 
 export const logCopyToClipboard = (
