+++ packages/headless/src/features/search/search-slice.test.ts
@@ -12,7 +12,8 @@ import {buildMockSearch} from '../../test/mock-search';
 import {buildMockSearchResponse} from '../../test/mock-search-response';
 import {buildMockSearchState} from '../../test/mock-search-state';
 import {createMockState} from '../../test/mock-state';
-import {AnalyticsType, makeAnalyticsAction} from '../analytics/analytics-utils';
+import {AnalyticsType} from '../analytics/analytics-utils';
+import {makeAnalyticsAction} from '../analytics/search-analytics-utils';
 import {applyDidYouMeanCorrection} from '../did-you-mean/did-you-mean-actions';
 import {logFacetShowMore} from '../facets/facet-set/facet-set-analytics-actions';
 import {logPageNext} from '../pagination/pagination-analytics-actions';
