+++ packages/headless/src/features/static-filter-set/static-filter-set-actions.ts
@@ -1,10 +1,7 @@
 import {createAction} from '@reduxjs/toolkit';
 import {validatePayload} from '../../utils/validate-payload';
-import {
-  AnalyticsType,
-  makeAnalyticsAction,
-  SearchAction,
-} from '../analytics/analytics-utils';
+import {AnalyticsType, SearchAction} from '../analytics/analytics-utils';
+import {makeAnalyticsAction} from '../analytics/search-analytics-utils';
 import {
   staticFilterIdSchema,
   staticFilterValueSchema,
