+++ packages/headless/src/features/static-filter-set/static-filter-set-insight-analytics-actions.ts
@@ -1,8 +1,8 @@
+import {AnalyticsType} from '../analytics/analytics-utils';
 import {
-  AnalyticsType,
   InsightAction,
   makeInsightAnalyticsAction,
-} from '../analytics/analytics-utils';
+} from '../analytics/insight-analytics-utils';
 import {getCaseContextAnalyticsMetadata} from '../case-context/case-context-state';
 import {LogStaticFilterToggleValueActionCreatorPayload} from './static-filter-set-actions';
