+++ packages/headless/src/features/triggers/trigger-analytics-actions.ts
@@ -3,11 +3,8 @@ import {
   requiredEmptyAllowedString,
   validatePayload,
 } from '../../utils/validate-payload';
-import {
-  AnalyticsType,
-  makeAnalyticsAction,
-  SearchAction,
-} from '../analytics/analytics-utils';
+import {AnalyticsType, SearchAction} from '../analytics/analytics-utils';
+import {makeAnalyticsAction} from '../analytics/search-analytics-utils';
 
 export interface LogUndoTriggerQueryActionCreatorPayload {
   /**
